+++ src/com_server/__init__.py
@@ -13,12 +13,11 @@ vers = sys.version_info
 if vers.major < 3 or (vers.major == 3 and vers.minor < 6):
     raise EnvironmentError("Python version >= 3.6 is required")
 
-from .api_server import ConnectionResource, RestApiHandler, EndpointExistsException
-from .constants import *
+from .api_server import (ConnectionResource, EndpointExistsException,
+                         RestApiHandler)
 from .base_connection import BaseConnection, ConnectException
 from .connection import Connection
-from .tools import all_ports, SendQueue, ReceiveQueue
-
-from .api_builtins import Builtins
+from .constants import *
+from .tools import ReceiveQueue, SendQueue, all_ports
 
 __version__ = "0.1.2"
